Make the update_path function more sensitive to .profile
Summary:
update_path() used to do a somewhat fragile check for /usr/local/sbin
being exported in a user's .bash_profile. However, if the user had a
preconfigured .profile or .bash_login script (and no .bash_profile),
this script would no longer be run since bash prefers to run
.bash_profile if it can find it. Now there is a more robust check in all
of the bash conf files and if /usr/local/sbin is not found in an export
it will be added to the file that's currently first on bash's exucution
list, or .profile if no such file exists.

Also, fixed a potential bug where the current path was being used for
the export command placed in the config file.

Test Plan:
Ran make with various combinations of .bash_profile, .bash_login, and
.profile files both with and without an export containing
/usr/local/sbin.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2262