fix command line tools download locations for 10.8
Summary: the old paths were... old, but also, the automated mount/installer command was probably breaking anyway. This whole section is fragile, but this at least attempts to make it less fragile

Test Plan: tested the bash commands on a new laptop running 10.8, but not on 10..7, but the DMGs are named in the similar style.

Reviewers: csilvers

Reviewed By: csilvers

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D3802