actually install hipchat
Summary:
previously, the script just opened the hipchat website. now it actually checks to see if hipchat is available and if not, will download it, copy it to the user's ~/Applications dir and then open it. It can be run multiple times, it will only install once.

because it uses `open` to 'discover' the location of hipchat, it will even work if somebody previously installed hipchat to their desktop or some other novel location.

It will not succeed if hipchat was installed to the trash.

Test Plan: http://dl.dropboxusercontent.com/u/406291/Screenshots/7l3R.png

Reviewers: csilvers, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D3817