Increase the per-shell limit for open file descriptors
Summary:
This change should solve the "too many open files" errors when running tests.
However the root issue may be that the test harness is leaking file
descriptors. I haven't looked into that.

On OS X 10.8.4, the default ulimit -n of 256 is the limiting factor, not the
kernel sysctls for maxfiles and maxfilesperproc.

Test Plan:
Tested the ulimit change's effect on tests by running the following commands:
  sudo sysctl -w kern.maxfiles=12288  # OS X default
  sudo sysctl -w kern.maxfilesperproc=1024  # OS X default
  ulimit -S -n 256  # OS X default
  make test  # fails
  ulimit -S -n 2560  # 10x the OS X default
  make test  # succeeds

Tested the addition to .bashrc.khan by running the following commands:
  ulimit -n  # => 256
  echo "ulimit -S -n 2560" >> ~/.zshrc  # I use zsh
  source ~/.zshrc
  ulimit -n  # => 2560

Reviewers: csilvers

Reviewed By: csilvers

CC: jace

Differential Revision: http://phabricator.khanacademy.org/D4007