no longer install nginx as part of mac dotfiles setup
Summary:
our use of nginx as a reverse proxy in front of dev_appserver and 'the filesystem' was originally set up as a result of dev_appserver's sluggishness at serving static files. that problem has pretty much abated since the release of
v2 of the dev server.

Test Plan: tested on fresh mac, nginx not installed.

Reviewers: chris, csilvers, desmond

Reviewed By: csilvers, desmond

Subscribers: desmond, csilvers, dylan

Differential Revision: http://phabricator.khanacademy.org/D8965