simplify osx command-line devtools install
Summary:
since most dotfiles running folks are on 10.9 and the devtools links change often, it's safer to point people to developer.apple.com for os installs that are more than a year old. i suspect any 10.8 holdouts on the devteam are probably also comfortable installing devtools.

for people who have the tools installed, we just mention that we think they're ok.

Test Plan: tested function in isolation on a new mac.

Reviewers: chris

Reviewed By: chris

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D9016