Add support for a new branch.review-parent-default config option.
Summary:
If this is set, then
   git review-branch foo
will branch foo off of the review-parent-default branch (which must
exist).  If it's not set, 'git review-branch foo' retains the current
behavior of branching foo off of the current active branch.

Either way, manually specifying the parent branch -- as the second
argument to review-branch -- takes precedence.

Test Plan:
Ran
   git rb stale-lock-msg
after putting in this code, while on the master branch, and it said:
   Switched to branch 'csilvers'
   Branch stale-lock-msg set up to track local branch csilvers by rebasing.

Reviewers: alpert

Reviewed By: alpert

Subscribers: desmond

Differential Revision: http://phabricator.khanacademy.org/D11152