Add a pre-push hook to go along with the pre-commit hook.
Summary:
This is to add extra insurance that you don't accidentally push to
master on github.  The idea is the commit hook keeps you from
committing the thing you'd push to master, and the push hook keeps you
from actually doing the push.

If, say, you're in the habit of doing 'git commit -n' -- naughty
naughty! -- this second check will save you from making a Big
Mistake.  Likewise if you're the kind of person who does:
   % git push origin mybranch:master

Test Plan:
I did an empty commit in the master branch (using commit -n) and then
did
   git push
and it failed with
   FATAL ERROR: You cannot push to the master branch on github.

I then went to the deploy branch and did
   git merge master   # to have some changes in it
   git push
and it succeeded.

Finally, I did
   git push origin deploy-test:master
and it took failed with
   FATAL ERROR: You cannot push to the master branch on github.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D11206