Totally revamp the way we handle .bashrc.
Summary:
I have spent some time groveling through the bash manpages, and am
clearer now on the difference between .bashrc and .bash_profile --
basically, one is used for login shells and one for non-login shells.

We want our code to run in both contexts, so I have moved everything
to .bash_profile, which I now source from .bashrc.  (I actually only
source .bash_profile.khan, to keep any existing .bash_profile stuff
out of it, since maybe the user *does* want that stuff to be
login-only.)

I also separate out .bash_profile -- which is bash-specific -- from
.profile -- which runs on any sh-compatible shell, including zsh.
This should make it easier for zsh users to do their setup.

Test Plan:
I created a new user on my machine, and ran
   make
in the khan-dotfiles directory.  It created all the files the way they
should be.  I then ran
   bash
   git rb
and saw that the git alias was picked up; that is, the git command
didn't say 'rb not found'.  I then ran
   bash --login
   git rb
and saw that it was picked up as well (it wasn't, before this change).

Reviewers: marcos, alpert

Reviewed By: alpert

Subscribers: ayman

Differential Revision: http://phabricator.khanacademy.org/D13071