The second half of 114cc6c8ec0162df52fc84404e6b3cf8d1cf03a6.
Summary:
I got rid of the 'git co master' but didn't get rid of the 'git co -'
at the end (that was the 'pop' to go along with the original git-co
'push').  Oops!

I also changed the commit message to be a bit more descriptive.

Test Plan:
Ran
   git update-submodule third-party/frankenserver
   git log
and saw it say:

    Substate for third_party/frankenserver-khansrc

    Test plan:

    None

Much nicer.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D13784