Make sure xcode-select is run when /usr/include is empty.
Summary:
Before, we were running xcode-select whenever gcc was missing, but not
otherwise.  But nowadays it seems that gcc can be around but we still
need to run xcode-select in order to populate /usr/include.  Shrug.
I dunno.

Test Plan:
Ran xcode-select --install manually for joe and had things work better
for him.

Reviewers: marcos

Reviewed By: marcos

Subscribers: joe, ben

Differential Revision: http://phabricator.khanacademy.org/D14258