add khan-linter bin to default PATH
Summary:
Now that khan-linter has ka-lint which can work from default path, add
it to the PATH like the other devtools.

In the future we should handle adding these paths in a more robust way
(in discussion), but for now just add it along with the other devtools
for consistency.

Test Plan: - source .profile.khan, ensure ka-lint is callable from PATH

Reviewers: csilvers, marcos

Reviewed By: csilvers, marcos

Subscribers: ethan

Differential Revision: https://phabricator.khanacademy.org/D18135