Remove outdated path munging for darwin
Summary:
We had some code here to make sure /usr/local/bin was in the path prior
to normal /usr/bin, as well as sbin.  This was in a darwin specific
context, and appears to have been for homebrew.

It was introduced in:
https://github.com/Khan/khan-dotfiles/commit/2bc3d0456e95499e211caf28aeae27cd04ef994e

However, this is a.) the default case on Darwin for a number of years
now, if you check /etc/paths you should be able to verify. b.) homebrew
checks for this case now both during installing and during `brew doctor`, so
it should be handled for us.

As for /usr/local/sbin, homebrew appears to no longer use it (in fact it
hasn't even been created on my system).  If at some point homebrew did
begin using it again, they would likely include in in their brew doctor
tests.

Test Plan:
- verify /usr/local/bin still in default PATH prior to /usr/bin after
  sourcing script
- wait and see if any weird problems creep up down the line

Reviewers: marcos, alpert, csilvers

Reviewed By: alpert, csilvers

Subscribers: ethan, alpert

Differential Revision: https://phabricator.khanacademy.org/D18137