Integrate ka-clone into the dotfiles setup process
Summary:
Uses ka-clone to clone all repositories, which means they have an
isolated local gitconfig with all KA settings. Rejiggers how we store
a global preference for user name and email so that it can be applied
to any future clones. As this removes the need for mucking about with
the global gitconfig, no longer include gitconfig.khan in the global
config (this will not remove it for existing users however [intentional],
even though pre-existing repositories that are checked out WILL be
upgraded via ka-clone --repair to have a nice local config.).

As a bonus, webapp and its dependencies are now an optional install,
which can be disabled via environment variable (useful to speed up
testing, but this should probably be made an option in the makefile
and exposed in the documentation in the future since its a common
request from the mobile team).

Test Plan:
test full install in MacOSX VM, verify completes, verify
repos have proper local gitconfig and git picks it up.

#devenvfixup

Reviewers: marcos, csilvers

Reviewed By: csilvers

Subscribers: mroth, ethan

Projects: #devenvfixup

Differential Revision: https://phabricator.khanacademy.org/D18609