remove outdated .ackrc
Summary:
after D18145 which removed the webapp specific paths from the global
ackrc, all that was left are some settings which in most cases represent
the defaults or have been covered in more recent versions of ack.  these
just had the function of overriding some global default behaviors of ack
in slightly nonstandard ways, so we should remove this entirely.

if we need anything specific to webapp, we can add that back in to its
repo .ackrc.

note: D18144 which re-added the specific webapp paths into that repo
still needs to be landed.

Test Plan: no testing required after review

Reviewers: marcos, ethan

Reviewed By: ethan

Subscribers: alpert

Projects: #devenvfixup

Differential Revision: https://phabricator.khanacademy.org/D18725