fix: don't update existing repos before ka-clone repair
It was a nice thought, but for upgrade path this can cause issues when
the user has their git repo in a dirty state.  We shouldn’t require
them to have it in a clean state, do don’t bother to git pull for now.

TODO to restore this behavior for devtools only, but make sure we
handle branches and dirty git states properly.

Auditors: ethan, csilvers