exclude more expected dotfiles by switching to whitelist
Summary:
There are other dotfiles which may appear in the root of the directory
that should not be symlinked.  Some are known parts of this repo
(.arclint for example), whereas others are autogenerated by OS operations
(such as .DS_Store on Darwin).

Symlinking .DS_Store can have particularly confusing results on a Mac,
because there are then two different directories the GUI is trying to
use the same file to store metadata regarding. (I’m pretty sure there
is an equivalent on many Linux window managers as well.)

We should definitely think about moving to the subdirectory plan
sometime in the near future for better organization, in the meantime, this
whitelist will have to be manually maintained (which is a reasonable tradeoff
for increased safety).

Test Plan: run in VM, verify .arc* files are not symlinked, but expected files are.

Reviewers: ethan, csilvers

Reviewed By: csilvers

Subscribers: alpert

Projects: #devenvfixup

Differential Revision: https://phabricator.khanacademy.org/D18831