dont modify resolv.conf
Summary:
FOR DISCUSSION

this is a nice hack for those who would want it, but it’s overly
prescriptive for us to modify without user intervention.  it’s also the
sort of one-liner we could just put in a document as a tip&trick for
those who want it.

also however, it actually doesn’t work properly for 90% of our
users(!), as MacOSX does not respect resolv.conf in recent versions and
instead uses the `networksetup` command to configure DNS resolution.

#devenvfixup

Test Plan: none

Reviewers: csilvers, ethan

Reviewed By: csilvers, ethan

Projects: #devenvfixup

Differential Revision: https://phabricator.khanacademy.org/D18991