Update dependencies and fix some bugs in linux-setup.sh
Summary:
This just fixes a few issues I encountered while running dev setup on linux.

A few dependencies were missing.  node-less in particular is needed because npm
installing it locally won't put `lessc` in a useful place.  Ubuntu also
installs the node binary as `/usr/bin/nodejs`, which will prevent most packages
from installing, so I linked it from `/usr/bin/node`.  If we don't think that's
a good idea, or want to do it differently, I can remove it.

The dpkg-redirect command failed for me; I'm not 100% sure but it seems like
the right thing to do is just ignore it if it fails.

Test Plan: Ran the setup script, it worked for me.

Reviewers: mroth, william, csilvers

Reviewed By: william, csilvers

Subscribers: mroth, alan

Differential Revision: https://phabricator.khanacademy.org/D20159