Install slack in linux dotfiles
Summary:
To avoid taking a side on the "where should `npm install -g` go" debate, we
install with sudo if we think we'll need it, and not otherwise.

(Note: this was originally written by @william, and modified and committed by
@benkraft.)

Test Plan:
Ran it, multiple times to check idempotency, with several choices of
npm prefix, and checked that it installed with or without sudo correctly.
Opened it from the Ubuntu menu to check that the `.desktop` file worked.

Reviewers: william, mroth, csilvers

Reviewed By: csilvers

Subscribers: william

Differential Revision: https://phabricator.khanacademy.org/D21694