Make sure users add virtualenv activation to their dotfiles
Summary:
If someone doesn't use `bash`, our bashrc won't actually do anything, and the
virtualenv won't get activated, which will cause later steps of dev setup to
fail.  For now we'll just error out and tell the user to fix it; hopefully
people with enough command-line experience to be using another shell will be
able to fix it themselves, or if not they can ping #1s-and-0s.

Test Plan:
* run the snippet in my shell, nothing happens
* set `SHELL=foo`, `VIRTUAL_ENV=""`, and define an `add_fatal_error` function,
  then run the snippet, and see it call the `add_fatal_error` function.

Reviewers: mroth, george, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D27388