Install timeout as an alias to gtimeout, and get rid of qcachegrind.
Summary:
We use `timeout` quite a bit in our deploy scripts and the like, but
on os x it's called gtimeout:
http://stackoverflow.com/questions/3504945/timeout-command-on-mac-os-x

This changes folks' configs so both commands work, in a minimal way.

I also got rid of the qcachegrind install when I was in the area,
since the chrome flamechart viewer works much better; we no longer
recommend qcachegrind to people.

Make sure gtimeout is installed before trying to alias it.

Test Plan:
Fingers crossed

Fingers crossed; it's a bit dangerous to do mac development on a linux
machine!  Luckily, I cut-and-paste all this stuff so it should be
relatively safe.

Reviewers: mroth

Reviewed By: mroth

Differential Revision: https://phabricator.khanacademy.org/D31367