Define KACLONE_BIN in mac-ios-setup.sh
Summary: `KACLONE_BIN` was not defined in mac-ios-setup.sh so running that file would crash on line 18 of shared-functions.sh. This fix is using the same code that is used in mac-android-setup.sh.

Test Plan:
run "sh ~/khan/devtools/khan-dotfiles/mac-ios-setup.sh"
It will not crash on line 18 of shared-functions.sh

Reviewers: hannah, bryan

Reviewed By: hannah, bryan

Subscribers: benkomalo, mroth, csilvers

Differential Revision: https://phabricator.khanacademy.org/D35778