Add a quick hooks-check to our common git aliases: `p` and `m`.
Summary:
The idea is that if your git-hooks aren't installed correctly, we'll
find out the next time you run `git p`.

I could have added the check for `git co` as well, but that command is
run *so* often I didn't want to slow it down even a tenth of a
second.  I could be convinced otherwise.  Checking hook-goodness once
a week or so is fine, and I figure people run `git p` at least that
often.

People who use IDE's that pull for them, and never run `git p`, won't
benefit from this (and probably are the ones who need it the most,
since who knows what their IDE will try to do for them).  Ah well.  If
there was another way to hook (ha) in these checks, I'd do it.

Test Plan: Ran `git p; echo $?` with success (and a 0 rc) after this change.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: mroth, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D39206