Add software necessary for working with protocol buffers to dotfiles
Summary:
See https://khanacademy.atlassian.net/browse/FASTBACK-172

Working with protocol buffers requires some extra software. We need the
protocol buffer compiler, which takes the protocol buffer definitions
and generates python and java code that we then use in webapp and
dataflow jobs. We also need a plugin for the compiler that allows us to
generate bigquery schemas too.

This commit ensures that these components are installed as part of
normal dev setup so that everyone is able to work with protocol
buffers. Because the compiler plugin is written in go and is
distributed only by source, we also need to install go as part of the
setup process.

Test Plan:
- in mac-setup.sh and linux-setup.sh, comment out all the steps but `install_protoc`
- `./mac-setup.sh`
- verify this completes successfully
- test on ubuntu 14.04 by:
  - `docker run -it -v ~/khan/devtools/khan-dotfiles:/khan-dotfiles ubuntu:14.04 /bin/bash`
  - install some components normally present in the desktop ubuntu version
    (or normally installed earlier in the process)
    - `apt-get update`
    - `apt-get install sudo software-properties-common apt-transport-https git wget unzip`
  - `cd /khan-dotfiles`
  - `./linux-setup.sh`
  - verify this completes successfully
- test on ubuntu 16.04 by the same process except using the docker command:
  `docker run -it -v ~/khan/devtools/khan-dotfiles:/khan-dotfiles ubuntu:16.04 /bin/bash`

Reviewers: nabil, csilvers

Reviewed By: csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D41241