Let `git co` work with phabricator revision-ids!
Summary:
Now, for repos that we have diff-tags set up (where phabricator
creates a tag on github every time we run `arc diff`), you can say
something like
    git co D1234
and it will check out the state of the repo in D1234.  (Technically
D1234 has a lot of states-of-the-repo, one for each call to `arc diff`
that was made to update that revision; we take the last one.)

It's not perfect: `git co -t D1234` won't work yet, you have to
specify the phabricator diff as the only arg.  But it's still pretty
useful!

Test Plan:
   git co _remove_badges
   git co -b test
   git co D43107
all worked as expected.  The first two were unchanged from the old
behavior, but the last one said:
   Note: checking out 'phabricator/diff/192259'.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D43293