Add some more startup info to dev setup scripts
Summary:
Now we only show the version warnings if you're on an unsupported
version -- I also added a warning for unsupported linux distros.
Additionally, we now warn if your login shell isn't bash.  I tried to
make all these checks pretty robust such that if they crash we assume
you're on an unsupported version.

Test Plan: Ran that part of the script on mac and linux

Reviewers: jacqueline, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D44277