Stop using system-site-packages on Mac
Summary:
The comment is apparently no longer accurate: I'm assuming
pyobjc is either compileable again or is bundling binary
deps; in any case I can install it just fine.  Meanwhile
`--system-site-packages` is causing us problems: it means
that the global installation of the python `protobuf` package
(via `brew install protobuf`) leaks into webapp's virtualenv,
which causes problems because it defines a `google` package,
incompatible with appengine's.  (This manifests strangely,
because we patch that conflict in `third_party/__init__.py`,
but if you haven't yet imported that, it doesn't help.
This causes `intl/i18n_lint_test.py` to fail, because it
shells out to a separate python process.)  Anyway, since
the flag is no longer needed, the easiest thing to do is
to remove it: it's just asking for this type of trouble.

Test Plan:
On a fresh macOS khan-dotfiles install (where
`tools/runtests.py intl/i18n_lint_test.py` fails):
- `rm -rf ~/.virtualenv/khan27 ~/Library/Caches/pip`
- in khan-dotfiles, `make` completes without error
- in webapp, `tools/runtests.py intl/i18n_lint_test.py` now passes,
  and `pip list` says pyobjc is installed.

Reviewers: csilvers, jacqueline

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D44413