Don't crash if an optional mac app fails to install
Summary:
The most common cause of this is that you already installed in manually
outside of brew -- in which case `brew ls` won't find it, but `brew
install` will fail.  In principle we could check for that, but knowing
how to check is very package-specific, so instead we just try, and warn
if we fail instead of crashing.

Test Plan: crossed fingers

Reviewers: Kai, jacqueline, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D44415