Tell pipenv not to use virtualenvs
In general I'm keeping pipenv setup in the repos that are using it so
far, but this one is really nice to have globally and doesn't really
cost anything.  It tells pipenv to ignore any virtualenv it finds itself
in, and create its own anyway.

Test plan: crossed fingers

Auditors: #devops