Install yarn as part of khan-dotfiles.
Summary:
We use it in webapp and we'll be using it more, so treat it the same
as npm.  We still use `npm` as our "bootstrap" system, though --
that's the one that's installed in a package-specific way.  `yarn` is
installed by npm.  I did not version yarn, the same way we don't
version npm; we always just get the most recent version.

I didn't change the way yarn is installed for mobile users since I
didn't understand how that works very well.  I do notice they use brew
instead of npm, but I don't know if there's a reason for that.

Test Plan: Manually ran the npm install comment.

Reviewers: benkraft, joshua, jared

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D50462