Log a much clearer fatal error on exit
Summary:
We keep having this problem where people run dotfiles, it crashes for
one reason or the other, and then they try to proceed and are surprised
when things don't work.  Now we make it super duper clear: if the script
exits anything but success, we log a loud warning.

Test Plan:
- comment out all the actual work to make it fast; `./setup.sh` works
  without error
- `./mac-setup.sh` fails (I'm on linux) and logs the error (and still
  exits 1)
- if I put `exit 1` inside `install_packages` in `./linux-setup.sh`, it,
  too, logs and exits 1

Reviewers: #devops, Kai, csilvers

Reviewed By: #devops, Kai, csilvers

Subscribers: Kai, csilvers

Differential Revision: https://phabricator.khanacademy.org/D50615