Update gcloud and install components in khan-dotfiles
Summary:
We're starting to need more components, and in practice many of them are
useful and you'll end up installing them anyway.  Now, we make sure you
are updated to the latest version as well as have the components
installed.  While I was at it, I bumped the version to match the new max
in webapp.

Test Plan: ran the commands manually, they were no-ops for me

Reviewers: #devops, csilvers

Reviewed By: #devops, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52503