Remove PhantomJS
Summary:
PhantomJS has been discontinued.  Homebrew gives this warning when you
attempt to install it via `brew cask`.

> phantomjs has been officially discontinued upstream.  It may stop
> working correctly (or at all) in recent versions of macOS.

After some discussion
[here](https://khanacademy.slack.com/archives/C02NMB1R5/p1555300186024400)
it was concluded that it's no longer a requirement to install this to
work on webapp.

Test Plan: Run `git grep phantomjs` and ensure there are no occurrences of PhantomJS in this code base.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: #frontend-infrastructure, jeff, csilvers

Differential Revision: https://phabricator.khanacademy.org/D53491