WEB-757: Add install_watchman to mac-setup.sh and linux-setup.sh in khan-dotfiles
Summary:
JIRA: https://khanacademy.atlassian.net/browse/WEB-757

Watchman is a file which jest uses to speed up its operation.  This diff
updates our dotfiles to install it if it doesn't exist.  We were already
doing this as part of the mac-ios and mac-android dotfiles so I move the
function that was already there into mac-setup.sh. On Linux, there's no
apt package so I added a script to install it from source.

Test Plan: - benkraft tested it (thank you)

Reviewers: #frontend-infrastructure, benkraft

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D54355