Always start postgres server, not just when installing
Summary:
Some users may have stopped their postgres server, which would lead to
an error when we try to connect to it to see if the correct postgres
user is created. This patch checks if the server is started and boots it
up if not.

Test Plan:
I manually stopped postgres by using `brew services stop postgresql`,
then ran `make os-install` and saw that it connected and no error was
produced when the script tried to create a postgres user.

I ran it again and saw the 'already started' message.

Reviewers: benkraft, Kai, csilvers

Reviewed By: benkraft, Kai, csilvers

Subscribers: csilvers, jeff

Differential Revision: https://phabricator.khanacademy.org/D55120