Remove special handling of `gae-` prefix in which-deploy and friends
Summary:
Back when we just had one version, we made it so you didn't have to type
the `gae-` on your version.  But now that's really more of a part of the
version, and anyway it's easier to copy-paste or interpolate into git
commands if we include it.  So we do.

Test Plan: `git prev-deploy $(git which-deploy master)` returns a plausible result

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: cathy, yansong, #devops

Differential Revision: https://phabricator.khanacademy.org/D55306