Make sure we install the right versions of go and protoc.
Summary:
We are using the go1.12 google runtime now, so we should have go1.12
on our dev machines to match.

Also, we need an old version (3.4) of the protoc protocol-buffer
compiler because of compatibility with some protobuf plugins, so we
make sure of that too.

I also no longer install protoc-gen-bq-schema here because it's finally
moved to go packages, so we can/do install it the normal way, via
go.mod.

Test Plan:
On linux, I ran install_protoc and install_go manually.
On OS X I ran `make` on a newly installed machine, and then ran
   go version
   protoc --version
to make sure it had the right versions of each.  I'm not 100%
sure this handles every possible upgrade case, but at least it's
not terrible.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: adamgoforth, marksandstrom

Differential Revision: https://phabricator.khanacademy.org/D57112