Fix Linux protoc installation
Summary:
On machines where the dotfiles have previously been run, protoc 3.5 was
installed. This was recently changed to install 3.4, but the command to
move the include files to the correct location doesn't work if there's
a previous version of them on the machine already. This commit changes
the script to first clean up the old files (if they exist) and then
move the new files into place.

Test Plan:
- Ran on my dev machine, which already had protoc 3.5 installed.
- Ran on a fresh Ubuntu 18.04 installation.

Reviewers: csilvers, benkraft

Reviewed By: csilvers

Subscribers: marksandstrom, csilvers

Differential Revision: https://phabricator.khanacademy.org/D57146