Install the required version of protoc on macOS
Summary:
This change updates the mac setup so that the required version of protoc
is install from a zip archive on github. If a homebrew version of protoc
is installed that isn't the current version, we uninstall it first.

The installation of protoc is the same for linux and mac, so the code is
shared.

Test Plan:
Remove my manually installed version of protoc. Run `make` without these
changes to install the latest homebrew version. After the changes, run
`make` and verify that the homebrew version is uninstalled and the
custom version is installed instead. Run `make` one last time to verify
that "protoc already installed".

Reviewers: csilvers, adamgoforth, benkraft

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D57167