Re-Apply the dotfile changes from last week
Summary:
Last week I made a number of changes to dotfiles, but apparently did something wrong with the phabricator part of the process, and only ended up committing a small subset to master.

Here attempting to fix that by cherry-picking the full change set and resolving merge conflicts.

https://github.com/Khan/khan-dotfiles/commit/8fbd32ef28abbc7e71168c3dce3c028f3e4664c1

Test Plan:
 # uninstall java8, wget, and go@1.12
# Rerun make, no errors (possibly need to rerun in new terminal for go)
# rerun in new terminal if needed
# rerun make, no errors, no new installs

Reviewers: Kai, csilvers

Reviewed By: csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D57921