Handle updating MacOS command line tools
Summary:
Re-running khan-dotfiles on an already existing setup (with existing but presumably out-of-date command line tools) was not properly handling the case of when our MacOS is at 10.14 but the command line tools are not. This was erroring out with `/Library/Developer/CommandLineTools/Packages/macOS_SDK_headers_for_macOS_10.14.pkg` not found.

Unfortunately seems there's no good way to upgrade from the command line, the recommended way on forums is to just remove the old version of the tools and reinstall via xcode. But at least let's make it slightly easier on our devs by restarting the setup script again for them. We put a note to wait before Pressing enter. If folks don't follow that advise, it's not terrible, the script will just loop back to the same place (but recognize that xcode-select --install is already happening, so it won't start it again). May be a little confusing for devs if they don't read that, but so is crashing the dotfiles!

Issue: none

Test Plan: Tested it on my local machine by renaming my SDK headers pkg to 10.13.pkg. It removed and then installed the correct one.

Reviewers: Kai, davidbraley, csilvers

Reviewed By: csilvers

Subscribers: kphilip, csilvers

Differential Revision: https://phabricator.khanacademy.org/D58691