Install Go 1.13 instead of 1.12 by default
Summary:
We're in the process of upgrading -- 1.13 is now supported on appengine
and has some nice new features.

Depends on D59736, which fixes a test currently failing on 1.13.  After
this and that both land, I'll ask folks to upgrade when they get a
chance; although 1.12 should continue to work for a little while
(probably until we add incompatible features or bump the version in
go.mod).

Issue: https://khanacademy.atlassian.net/browse/INFRA-3847

Test Plan: crossed fingers, ran the linux command, @Kai is running the mac command

Reviewers: #devops, O4 go-vernors, csilvers, Kai

Reviewed By: #devops, Kai

Subscribers: Kai

Differential Revision: https://phabricator.khanacademy.org/D59738