Also bump our check of whether to upgrade Go at all
Silly me, I thought bumping the version number in two places -- where we
upgrade it on Mac and Linux -- would be enough.  But it's not: we also
have to bump it where we decide to upgrade.  Now we will.

Note that 1.13 isn't strictly required yet, but we'll probably start
making use of 1.13 features very soon, at which point it will be, and
it's good for people to match prod anyway.

Issue: https://khanacademy.atlassian.net/browse/INFRA-3847

Test plan: crossed fingers

Auditors: csilvers, #devops