Add a script to rebuild your virtualenv
Summary:
I realized this is a common fix to dev environment problems, but the
usual instructions to do it are to delete the virtualenv, then re-run
khan-dotfiles, which can take a while.  But it's a lot easier than that.
Now we have a script, linked up to `make virtualenv`, to just redo the
virtualenv, assuming you already have python, webapp, etc.

Test Plan: ran the script, then `make serve`

Reviewers: #devops, csilvers, Kai

Reviewed By: #devops, csilvers, Kai

Subscribers: Kai, mattdunnrankin, csilvers

Differential Revision: https://phabricator.khanacademy.org/D60700