Fix handling of OSX gcc headers, git pcre. Add node version warning.
Summary:
Fixes the step to make sure git has PCRE2, which previously depeneded on a now unsupported flag. By just enforcing we're using git 2.20 or later, we're gauranteed to have pcre2.
Also adds a warning if the version of node isn't `v10.x.x`. We don't want to actually force the use of that version explicitly, there are valid reasons to use other versions, but we want to at least warn that it should be used.
Unfortunately, the problem with stdio.h on OSX 10.14 will have to persist, as the number of permutations to solve for is excessivly larger than the number of incidents of such problems we've seen. Additionally, our mechanisms for decting the problem are rudementary and potentially incorrect as well. Instead we detect (as best we can) and warn about a potential problem, advising the user to contact @dev-support if it turns out to be actually problematic.

Issue: INFRA-4588

Test Plan:
Tested in a vm running OSX10.14 with various versions of Xcode available or not.
Tested in a vm running OSX10.14 and 10.15 with various versions of git installed.
Tested running with various versions of node both installed/uninstalled and linked/unlinked.
Beyond that - fingers crossed.

Reviewers: csilvers, #devops

Reviewed By: csilvers, #devops

Subscribers: benkraft, samuel, csilvers

Differential Revision: https://phabricator.khanacademy.org/D62942