Better detection of pip, and remove brew doctor
Summary:
One person had a problem where `which pip` returned a location, but the command at that location was either not executable or corrupt. Attempt to mitigate that by running `pip --version` as well.
There was also some confusion around the output of 'brew doctor', and it's rarely if ever helpful, so removed it.

Issue: INFRA-2785

Test Plan: run `make install` on mac and ubuntu

Reviewers: #devops, avidal, benkraft, csilvers

Reviewed By: #devops, avidal, benkraft, csilvers

Subscribers: benkraft, avidal, csilvers

Differential Revision: https://phabricator.khanacademy.org/D63624